requirements.txt
flask<3
flask-mysql
boto3
requests
pyjwt[crypto]
sqids

python-dotenv

gunicorn[gevent]
